feat: enhance file download process in useDiagnose hook
- Refactored the download functionality in the useDiagnose hook to utilize a new utility function for file downloads.
- Improved error handling and user notifications for download failures.
- Added a new file utility function to streamline the download process, ensuring a better user experience.
