+++ frontend/webapp/utils/functions/index.ts
@@ -1,2 +1,3 @@
 export * from './destinations';
+export * from './files';
 export * from './sources';
